+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+# CHANGELOG
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+- Initial release
+- Add [TapWrapper] widget